+++ src/compiler/sema_types.c
@@ -307,7 +307,7 @@ static bool sema_resolve_type_identifier(SemaContext *context, TypeInfo *type_in
 
 
 
-// $typeof(...)
+// $Typeof(...)
 INLINE bool sema_resolve_typeof(SemaContext *context, TypeInfo *type_info)
 {
 	Expr *expr = type_info->unresolved_type_expr;
@@ -396,9 +396,9 @@ INLINE bool sema_resolve_typefrom(SemaContext *context, TypeInfo *type_info, Res
 			type_info->type = info->type;
 			return true;
 		case STORAGE_WILDCARD:
-			RETURN_SEMA_ERROR(expr, "$typefrom failed to resolve \"%.*s\" to a definite type.", (int)len, bytes);
+			RETURN_SEMA_ERROR(expr, "$Typefrom failed to resolve \"%.*s\" to a definite type.", (int)len, bytes);
 		case STORAGE_COMPILE_TIME:
-			RETURN_SEMA_ERROR(expr, "$typefrom does not support compile-time types.");
+			RETURN_SEMA_ERROR(expr, "$Typefrom does not support compile-time types.");
 	}
 	UNREACHABLE
 }